configs: proficube: Use distro_bootcmd approach
Finding and loading kernel and devicetree images can be implemented in
numerous ways. Although we won't support any "usual" Linux distribution
on this board, it makes sense to use a common approach on all of our
boards, at least for new boards. For ttCore we already used the "Generic
Distro Configuration Concept" which is documented in
'doc/README.distro'. The idea is to load kernel image etc. from the
rootfs and let the bootloader find and choose a bootable "disk". This
approach is both flexible, and generic.