board: proficube: Finish FPGA loading
This completes the SPI variant and adds the old GPIO based algorithm as
well. Some more things are configurable in the 'proficube_fpga.h' header
file now, but can stay as they are for later.

Speed with SPI is significantly faster than with GPIO. In fact the FPGA
is configured almost instantaneously now!

Support for SPI without U-Boot driver model was removed, we don't need
that anyway.