board: proficube: Configure PA21 as clock output for FPGA
On isNet Lite and v0.9 of the ProfiCube hardware we used the 50 MHz
clock between ethernet PHY and MAC as clock for FPGA, which was a bad
idea for different reasons. Now the FPGA should get its clock from
somewhere else and we opted to let the SoC generate that clock. We used
one of three configurable clock outputs the SAMA5D2 offers and select
the main clock (24 MHz) without prescaler, which is well suited for the
PROFIBUS signals the FPGA has to generate.

This approach is quick, dirty and hardcoded for now. It does use neither
DTS nor the U-Boot clock framework (and also not pinctrl). Leave all
those for later.

NOTE	This needs at least hardware revision v1.0. (Although it should
	not hurt on v0.9 where that SoC ball simply is not connected.)
	On v0.9 you'll still need to configure the FPGA with an image
	for 50 MHz clock. From v1.0 onwards the FPGA has to be
	configured with a suitable image for 24 MHz clock!