board: proficube: Bring SMC timings in line with Linux
After implementing the "backplane" test in U-Boot we noticed the test
failing in U-Boot with bitflips after some megabytes of transmitted
data, but not failing in Linux. This might be due to faster timings of
the external memory interface in U-Boot. In Linux those timings are set
in device tree in nanoseconds, the drivers take care of computing the
right values for the registers. We now just take what Linux set, and set
it in U-Boot as well. The test don't fails anymore.

Fixes: e70192b9d73b ("board: proficube: Optimize SMC timings")
Acked-by: Jens Bertram <jbe@thorsis.com>