arm: dts: at91: muxgate: Add nodes for 74xx595 and more LEDs
The board has four two pin dual color LEDs (red/green) connected to the
eight pins of a 74xx595 shift register connected to the SPI0 bus of the
SoC.

In .dts we added the necessary nodes and let the spi controller handle
CS0 natively (not via gpio) now.

The actual gpio driver for the 74xx595 circuit was activated in
defconfig.