board: muxgate: Add tt commands
We already had the commands 'tt hello' and 'tt ledtest' on the ProfiCube
board. The LED test is valuable for production, so we want that for
MUXgate, too. However the LEDs on MUXgate are connected a little
different, still through a 74xx595, but that one is connected to the SoC
on MUXgate, not to the FPGA like on ProfiCube.

Moved the common parts from proficube to a new folder
'board/thorsis/common' and implemented the muxgate part against the
U-Boot led API.