board: proficube: Add support for 64 MiB SiP variant (D5M)
The linux kernel sees the amount of memory reported in the memory node
of the device tree passed on boot. U-Boot changes or fixes that memory
size when calling bootm or bootz based on what was set as gd->ram_size
before. This way we can do the whole SoC and RAM detection in U-Boot
just based on the CPU registers which tell us the SiP variant. The
memory set in U-Boot will be set through fixup, regardless what's in the
memory node in dts loaded from /boot partition.