board: ncl: Migrate button config
With 371244cb19 ("Makefile: Give a build error if ad-hoc CONFIG options
are added") ad-hoc options in include/configs/[board].h are not allowed
anymore.

Since proper button support in U-Boot is not supported yet in v2016.11,
we just move these options, one to the board config file and another one
to the C code (because that won't be changed anyways).

For: v2016.11