board: ncl: Remove an ancient (non-)option
With 371244cb19 ("Makefile: Give a build error if ad-hoc CONFIG options
are added") ad-hoc options in include/configs/[board].h are not allowed
anymore.

The option removed here was introduced when first adding board support
back in 2010.  It is always set, in fact it makes no sense to not set
it, because it initializes board related things.  So just don't make it
optional anymore.

For: v2016.11