board: ncl: Move CONFIG_USER_LED to board code
With 371244cb19 ("Makefile: Give a build error if ad-hoc CONFIG options
are added") ad-hoc options in include/configs/[board].h are not allowed
anymore.

U-Boot has a CONFIG_CMD_LED already in v2016.11, but it's not enabled
(yet), because it uses Device Model (DM) and so on.  At this version
Atmel boards still use that old stuff without (GPIO) LEDs configured
through Device Tree (DT), so we postpone it here too, and only fix the
build error.

For: v2016.11