configs: ncl: Move initial image load address
The previous address where at91bootstrap loaded the U-Boot image to was
2 MiB below top of RAM.  After the U-Boot image itself grew too large
with the upgrade to v2017.07 there was an overlap between the image in
memory and the designated post relocate malloc area.  To avoid this
overlap we move the load address to 4 MiB below top of RAM now.  This
should not hurt usual operation, because relocation still happens
(correctly).

Note 1: Although not knowing exactly, we suspect we had a similar
problem on Vera2 and ProfiCube when adding support for boards with fewer
SDRAM, where changing that load address also solved the hangs.

Note 2: This probably does not happen on at91sam9g20ek board (which the
isNet Lite is largely based on), because that one has 64 MiB of SDRAM
instead of the 32 MiB we have on isNet Lite.