board: thorsis: Imply DISTRO_DEFAULTS but don't set it
We use the distro_defaults approach to boot and eventually call 'run
distro_bootcmd' on boot for our sama5d27 based boards.  However we don't
select DISTRO_DEFAULTS in Kconfig but only use a subset of options set
by that.

On ProfiCube (and later MUXgate) we set the env variable 'bootcmd'
explicitly in scripts when bootstrapping those boards in production.
For Vera2 the variable was part of default env up to U-Boot v2022.01,
but not anymore with U-Boot v2022.04 after 970bf8603b87 ("Convert
CONFIG_USE_BOOTCOMMAND et al to Kconfig").  It was not (over)written by
scripts however.  Thus when building Vera2 from defconfig (or the
actually used config in BSP) we could not boot anymore after upgrading
to v2022.04, which hit us hard in BSP-tt-generic.

This change uses a weak dependency on DISTRO_DEFAULTS through 'imply'
now to indicate we want to use that approach, but explicitly disable it
in defconfig and disable things we probably never need on those targets.
The CONFIG_BOOTCOMMAND is set through defconfig to the default implied
by DISTRO_DEFAULTS now, which lets all our boards behave like before
v2022.04.

(It does not matter to have that default on ProfiCube et al. because
it's simply overwritten then.)

Fixes: 970bf8603b87 ("Convert CONFIG_USE_BOOTCOMMAND et al to Kconfig")