fixup! Convert CONFIG_USE_BOOTCOMMAND et al to Kconfig
We don't need that default bootcommand on ncl.  The bootcmd env variable
is populated by "Vortest" and later by startup u-boot env, but both with
content not usable in a default boot command.

The original change was with v2022.01-rc3-132-g970bf8603b but the
adaption slipped through. (Hopefully won't happen in the future. We have
better tooling to detect such changes now.)