fixup! ARM: dts: at91: sama5: Align with Linux Devicetree
Easy thing for ProfiCube and MuxGate.  We used the opportunity to move
the things to referenced nodes (for all affected boards) right away.

For Vera2 the gpio-hog thing is tricky, moving the pinctrl references to
the hog node seemed like the cleanest idea, not tested yet.  The
no-1-8-v was added, we have that in Linux dts in the same way.

Maybe we should see if we can synchronize Linux and U-Boot .dts someday.

Fixes: v2022.10-rc3-253-g2df729e96d ("ARM: dts: at91: sama5: Align with Linux Devicetree")
For: v2023.01