board: thorsis: hartcube: Complete LED support
* Move things out of .dts which are most probably U-Boot specific anyways
* Settle on the led to show boot status for now
* Init all the non-debug LEDs so that status led will actually light up
* Implement the tt ledtest

Note: the labels or not fixed yet, because we did not assign any
functions in "main" dts for linux yet.  Those might change again in the
future, but since they are defined in u-boot .dtsi now, everything we
have to do is keep them synced with the tt ledtest code for HARTcube.