ARM: dts: at91: hartcube: Add dbgu and move sdhci
sam9x60.dtsi differ quite a bit between Linux (upstream) and U-Boot.
The latter enables some common nodes, which is not done in Linux.

To minimize differences in at91-hartcube.dts between Linux and U-Boot
version we add the dbgu nodes here which are needed in Linux, although
already defined in U-Boot's sam9x60.dtsi.

For the sdhci nodes: those are disabled in Linux already, we only need
to disable those in U-Boot, because those are some of the common enabled
nodes in sam9x60.dtsi here.