fpga: Introduce new driver for passive serial configuration on Altera
Passive serial configuration over SPI was previously done by board code
implementing the board specific functions and hooking that into the
ancient FPGA framework.  On modern boards it should be sufficient to
just define an fpga subnode for the spi node in dts and let this generic
driver do the rest.

The 'cookie' integer in the old FPGA interfaces is used to communicate
SPI interface details to the board specific functions already, at least
there are some macros for that in altera.h which are used by some
boards.  Here we "misuse" that cookie to pass a pointer to the udevice
to the hook functions, which allows us to get SPI and GPIO interfaces
through that device pointer.

Device tree bindings are the same as on Linux.  The family is encoded in
the compatible string, the device size can be set in device tree as
well.