fpga: cyclon2: Send dummy bytes without toggling CS in between
The AN 006 states:

> Important: To ensure a successful configuration, the microprocessor
> must continue to supply configuration clock to the Trion® FPGA for
> at least 100 cycles after sending the last configuration data.

Previously we split that into two transfers, each toggling CS line, so
we had a CS de-assert and assert in between the data and the dummy
bytes.  Did not work on hartcube with the new atmel-quadspi driver.  Now
we read memory we did not allocate, but it's only reading and there's no
virtual memory and no kernel around who would forbid this, so just read
that garbage to keep the clock alive.