board: thorsis: proficube: Migrate to generic FPGA configuration driver
After defining everything necessary in DTS, do the same as for HartCube.

Previously the board code decided which method to use depending on some
environment variable containing the FPGA image to be loaded.  This would
be highly impractical now when everything is defined in DT, so the new
approach is to build different u-boot binaries with different (control)
DT for the different hardware variants/revisions.

Note: this currently only applies to U-Boot which needs to distinguish
the FPGA for configuring.  The DT for Linux can be the same as before,
because in our case Linux considers/requires the FPGA as configured
already and does not care anymore about the different parameters for the
SPI peripheral device.  That is why CONFIG_DEFAULT_FDT_FILE stays the
same.

(Making this distinction in Linux would mean to introduce multiple
device tree files over there with all kinds of backward and forward
compatibility questions/issues.  We won't open that can of worms until
really required.)